Project Generator Responsive Bug Fix
The Duplication of the quick project result container for mobile caused and issue with the creation and display of a quick project on desktop. With some class and id changes along with amendments to the javascript the bug is now fixed and the function works as intended on mobile and desktop devices.